chore: move types to separate files
