+++ src/services/user-preferences/models/definitions/user-preference-definition.ts
@@ -1,7 +1,6 @@
 import { type UserPreferenceScopeType } from './user-preference-scope-type'
 
-export interface UserPreferenceDefinition {
-  isOptedInByDefault: boolean
+export type UserPreferenceDefinition<T> = {
   allowedScope: UserPreferenceScopeType
-  defaultData: Record<string, unknown>
+  defaultValue: T
 }
