+++ src/services/user-preferences/models/definitions/user-preference-definitions.ts
@@ -1,5 +1,5 @@
 import { type UserPreferenceDefinition } from './user-preference-definition'
 
-export type UserPreferenceDefinitions<UserPreferenceId extends PropertyKey> = {
-  [Id in UserPreferenceId]?: UserPreferenceDefinition
+export type UserPreferenceDefinitions<T> = {
+  [P in keyof T]: UserPreferenceDefinition<T[P]>
 }
