+++ src/services/user-preferences/models/definitions/user-preference-per-scope.ts
@@ -1,0 +1,5 @@
+import { type UserPreferenceScope } from '../storage-models'
+
+export type UserPreferencesPerScope<T> = {
+  [K in UserPreferenceScope]?: T
+}
